requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
psycopg[binary]==3.2.9
PyMuPDF==1.26.3
python-multipart==0.0.20
pytest==8.4.1
pytest-cov==6.2.1
httpx==0.28.1
